Pinned common lib to latest version, fixed tests.
A number of tests were broken by a tightening of the schema for TRT XMLs, which was done as a hotfix in Tier 3 support and merged into development. Most of the fixes were just updating the sample TRTs used as test inputs to be compliant with the new schema. In a few cases, the tests themselves had to be updated to conform with the new behavior.